requirements.txt
Pillow==5.3.0
pytz
RPi.GPIO
matplotlib